+++ libft/src_libft/ft_memmove.c
@@ -24,11 +24,11 @@
 
 #include "../include/libft.h"
 
-void	*ft_memmove(void *dest, const void *src, size_t n)
+void	*ft_memmove(void *dest, const void *src, int n)
 {
 	char	*t_src;
 	char	*t_dest;
-	size_t	i;
+	int	i;
 
 	t_src = (char *)src;
 	t_dest = (char *)dest;
