+++ src/ft_parse_args.c
@@ -6,7 +6,7 @@
 /*   By: xenia <xenia@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/01 14:45:40 by xenia             #+#    #+#             */
-/*   Updated: 2024/09/21 12:21:17 by xenia            ###   ########.fr       */
+/*   Updated: 2024/09/21 12:29:22 by xenia            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ static void	ft_init_limits(t_stack **stack, int v)
 static void	ft_parse_long(t_stack **st, char *arg)
 {
 	t_dlist	*new;
-	size_t	i;
+	int	i;
 
 	if (!arg || !*arg || ft_just_space(arg))
 	{
